backend/requirements/base.txt
# Django core
Django>=5.1,<5.2
djangorestframework>=3.15
django-cors-headers>=4.4
django-filter>=24

# Authentication
djangorestframework-simplejwt>=5.3

# Database
psycopg[binary]>=3.2

# Cache & message broker
redis>=5.0

# WebSockets / async
channels>=4.1
channels-redis>=4.2
daphne>=4.1

# Background tasks
celery>=5.4
django-celery-beat>=2.7
django-celery-results>=2.5

# QR codes
qrcode[pil]>=7.4
Pillow>=10

# Utilities
python-decouple>=3.8
django-extensions>=3.2
